move wip new v0.6 schema to etl, because it is partly used there
