+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/oshdb/OSMMemberRelation.java
@@ -1,8 +1,6 @@
-package org.heigit.bigspatialdata.oshdb.osm2.impl;
+package org.heigit.bigspatialdata.oshdb.tool.importer.transform.oshdb;
 
-import org.heigit.bigspatialdata.oshdb.osh2.OSHEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
-import org.heigit.bigspatialdata.oshdb.osm2.OSMMember;
 
 public class OSMMemberRelation implements OSMMember {
