+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/oshdb/OSMMemberWay.java
@@ -1,15 +1,17 @@
-package org.heigit.bigspatialdata.oshdb.osm2.impl;
+package org.heigit.bigspatialdata.oshdb.tool.importer.transform.oshdb;
 
-import org.heigit.bigspatialdata.oshdb.osh2.OSHEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
-import org.heigit.bigspatialdata.oshdb.osm2.OSMMember;
 
-public class OSMMemberWayIdOnly implements OSMMember{
+public class OSMMemberWay implements OSMMember{
 
   private final long id;
-    
-  public OSMMemberWayIdOnly(long id) {
+  private final OSHNode node;
+  
+  
+  
+  public OSMMemberWay(long id, OSHNode node) {
     this.id = id;
+    this.node = node;
   }
 
   @Override
@@ -28,11 +30,10 @@ public class OSMMemberWayIdOnly implements OSMMember{
   }
 
   @Override
-  public OSHEntity getEntity() {
-    return null;
+  public OSHNode getEntity() {
+    return node;
   }
 
-  
   @Override
   public String toString() {
     return asString();
