+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/oshdb/OSMMemberWayIdOnly.java
@@ -1,19 +1,13 @@
-package org.heigit.bigspatialdata.oshdb.osm2.impl;
+package org.heigit.bigspatialdata.oshdb.tool.importer.transform.oshdb;
 
-import org.heigit.bigspatialdata.oshdb.osh2.OSHNode;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
-import org.heigit.bigspatialdata.oshdb.osm2.OSMMember;
 
-public class OSMMemberWay implements OSMMember{
+public class OSMMemberWayIdOnly implements OSMMember{
 
   private final long id;
-  private final OSHNode node;
-  
-  
-  
-  public OSMMemberWay(long id, OSHNode node) {
+    
+  public OSMMemberWayIdOnly(long id) {
     this.id = id;
-    this.node = node;
   }
 
   @Override
@@ -32,10 +26,11 @@ public class OSMMemberWay implements OSMMember{
   }
 
   @Override
-  public OSHNode getEntity() {
-    return node;
+  public OSHEntity getEntity() {
+    return null;
   }
 
+  
   @Override
   public String toString() {
     return asString();
