+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/oshdb/OSMNode.java
@@ -1,4 +1,4 @@
-package org.heigit.bigspatialdata.oshdb.osm2;
+package org.heigit.bigspatialdata.oshdb.tool.importer.transform.oshdb;
 
 public interface OSMNode extends OSMEntity{
