+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/reader/TransfromNodeReaders.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.TreeSet;
-import org.heigit.bigspatialdata.oshdb.tool.importer.osh.TransformOSHNode;
+import org.heigit.bigspatialdata.oshdb.tool.importer.transform.oshdb.TransformOSHNode;
 import org.heigit.bigspatialdata.oshdb.tool.importer.util.ZGrid;
 
 public class TransfromNodeReaders {
